% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_account.R
\name{get_account}
\alias{credential_report}
\alias{get_account}
\title{Get Account Details}
\usage{
get_account(...)

credential_report(...)
}
\arguments{
\item{\dots}{Additional arguments passed to \code{\link{iamHTTP}}.}
}
\value{
A list containing various account details.
}
\description{
Retrieve IAM Account Details. This is useful as a \dQuote{hello world!} test.
}
\details{
\code{get_account} returns a list of account details. \code{credential_report} generates and/or retrieves a credential report.
}

