% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login_profile.R
\name{create_login}
\alias{create_login}
\alias{delete_login}
\alias{get_login}
\alias{update_login}
\title{Instance Profiles}
\usage{
create_login(user, password, reset, ...)

update_login(user, password, reset, ...)

delete_login(user, ...)

get_login(user, ...)
}
\description{
Coming soon\dots
}

