\name{aws.iam-package}
\alias{aws.iam-package}
\alias{aws.iam}
\docType{package}
\title{AWS IAM Client Package}
\description{A simple client package for the Amazon Web Services (AWS) Identity and Access Management (IAM) API.}
\details{}
\author{Thomas J. Leeper <thosjleeper@gmail.com>}
%\references{}
\keyword{ package }
%\seealso{}
%\examples{}
