% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_account.R
\name{get_account}
\alias{auth_details}
\alias{credential_report}
\alias{get_account}
\title{Get Account Details}
\usage{
get_account(...)

credential_report(...)

auth_details(type, n, marker, ...)
}
\arguments{
\item{type}{An optional character string specifying one or more types of reports to return.}

\item{n}{An integer specifying the number of responses to return.}

\item{marker}{A character string specifying a marker (from a previous response) to use in paginating results}

\item{\dots}{Additional arguments passed to \code{\link{iamHTTP}}.}

\item{\dots}{Additional arguments passed to \code{\link{iamHTTP}}.}
}
\value{
A list containing various account details.
}
\description{
Retrieve IAM Account Details. This is useful as a \dQuote{hello world!} test.
}
\details{
\code{get_account} returns a list of account details. \code{credential_report} generates and/or retrieves a credential report. \code{auth_details} returns a list of group, user, role, and policy details.
}

