% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http.R
\name{iamHTTP}
\alias{iamHTTP}
\title{Workhorse API Query Function}
\usage{
iamHTTP(query, verb = "GET", body = "", version = "2010-05-08",
  region = Sys.getenv("AWS_DEFAULT_REGION", "us-east-1"),
  key = Sys.getenv("AWS_ACCESS_KEY_ID"),
  secret = Sys.getenv("AWS_SECRET_ACCESS_KEY"), ...)
}
\arguments{
\item{query}{A named list specifying query arguments.}

\item{verb}{A character string specifying an HTTP verb. Either \dQuote{GET} or \dQuote{POST}.}

\item{body}{A character string specifying a request body (if \code{verb = "POST"}).}

\item{version}{A character string specifying an API version. Default is \dQuote{2010-05-08}.}

\item{region}{A character string specifying an AWS region. The default is drawn from environment variable \env{AWS_DEFAULT_REGION}.}

\item{key}{A character string specifying an AWS Access Key. The default is drawn from environment variable \env{AWS_ACCESS_KEY_ID}.}

\item{secret}{A character string specifying an AWS Secret Key. The default is drawn from environment variable \env{AWS_SECRET_ACCESS_KEY}.}

\item{dots}{Additional arguments passed to \code{\link[httr]{GET}} or \code{\link[httr]{POST}}}
}
\description{
This is the low-level API querying function for IAM. Users do not need to use this directly.
}

