% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instance_profile.R
\name{create_profile}
\alias{create_profile}
\alias{delete_profile}
\alias{get_profile}
\alias{list_profiles}
\title{Instance Profiles}
\usage{
create_profile(profile, path, ...)

delete_profile(profile, ...)

get_profile(profile, ...)

list_profiles(role, n, marker, prefix, ...)
}
\description{
Coming soon\dots
}
\references{
\href{http://docs.aws.amazon.com/IAM/latest/UserGuide/id_roles_use_switch-role-ec2_instance-profiles.html}{About Instance Profiles}
}

